Refactor node inventory helpers into inventory module
