+++ custom_components/termoweb/backend/termoweb_ws.py
@@ -51,10 +51,13 @@ from custom_components.termoweb.installation import (
     InstallationSnapshot,
     ensure_snapshot,
 )
-from custom_components.termoweb.inventory import normalize_node_addr, normalize_node_type
+from custom_components.termoweb.inventory import (
+    build_node_inventory as _build_node_inventory,
+    normalize_node_addr,
+    normalize_node_type,
+)
 from custom_components.termoweb.nodes import (
     addresses_by_node_type as _addresses_by_node_type,
-    build_node_inventory as _build_node_inventory,
     collect_heater_sample_addresses,
     heater_sample_subscription_targets,
     normalize_heater_addresses,
