+++ custom_components/termoweb/nodes.py
@@ -2,18 +2,14 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable, Iterable, Mapping, MutableMapping
+from collections.abc import Iterable, Mapping, MutableMapping
 import logging
 from typing import Any, cast
 
 from .const import DOMAIN
 from .inventory import (
-    AccumulatorNode,
-    HeaterNode,
-    NODE_CLASS_BY_TYPE,
     Node,
-    PowerMonitorNode,
-    ThermostatNode,
+    build_node_inventory,
     normalize_node_addr,
     normalize_node_type,
 )
@@ -22,233 +18,6 @@ _LOGGER = logging.getLogger(__name__)
 
 
 HEATER_NODE_TYPES: frozenset[str] = frozenset({"htr", "acm"})
-
-_NODE_SECTION_IGNORE_KEYS = frozenset(
-    {"dev_id", "name", "raw", "connected", "nodes", "nodes_by_type"}
-)
-
-_SNAPSHOT_NAME_CANDIDATE_KEYS = (
-    "name",
-    "label",
-    "title",
-    "display_name",
-    "device_name",
-    "friendly_name",
-    "heater_name",
-    "alias",
-    "room",
-)
-
-
-def _existing_nodes_map(source: Mapping[str, Any] | None) -> dict[str, dict[str, Any]]:
-    """Return a mapping of node type sections extracted from ``source``."""
-
-    if not isinstance(source, Mapping):
-        return {}
-
-    sections: dict[str, dict[str, Any]] = {}
-
-    raw_existing = source.get("nodes_by_type")
-    if isinstance(raw_existing, Mapping):
-        for node_type, section in raw_existing.items():
-            if isinstance(section, Mapping):
-                sections[node_type] = dict(section)
-
-    for key, value in source.items():
-        if key in _NODE_SECTION_IGNORE_KEYS:
-            continue
-        if isinstance(value, Mapping):
-            sections.setdefault(key, dict(value))
-
-    return sections
-
-
-def _iter_snapshot_sections(
-    sections: Mapping[str, Any],
-    seen: set[tuple[str, str]],
-) -> Iterable[dict[str, Any]]:
-    """Yield node payloads derived from snapshot-style ``sections``."""
-
-    for node_type, payload in sections.items():
-        if not isinstance(node_type, str):
-            continue
-        normalized_type = node_type.strip()
-        if not normalized_type or not isinstance(payload, Mapping):
-            continue
-        for entry in _iter_snapshot_section(normalized_type, payload):
-            addr = entry.get("addr")
-            if not isinstance(addr, str):
-                continue
-            key = (normalized_type, addr)
-            if key in seen:
-                continue
-            seen.add(key)
-            yield entry
-
-
-def _iter_snapshot_section(
-    node_type: str, section: Mapping[str, Any]
-) -> Iterable[dict[str, Any]]:
-    """Yield node dictionaries for a single node type section."""
-
-    addresses = _collect_snapshot_addresses(section)
-    for addr in sorted(addresses):
-        entry: dict[str, Any] = {"type": node_type, "addr": addr}
-        name = _extract_snapshot_name(addresses[addr])
-        if name:
-            entry["name"] = name
-        yield entry
-
-
-def _collect_snapshot_addresses(
-    section: Mapping[str, Any]
-) -> dict[str, list[Mapping[str, Any]]]:
-    """Return mapping of addresses to candidate payloads from ``section``."""
-
-    addresses: dict[str, list[Mapping[str, Any]]] = {}
-
-    addrs = section.get("addrs")
-    if isinstance(addrs, (list, tuple, set)):
-        for candidate in addrs:
-            addr = normalize_node_addr(candidate)
-            if addr:
-                addresses.setdefault(addr, [])
-
-    for value in section.values():
-        if not isinstance(value, Mapping):
-            continue
-        for addr_key, payload in value.items():
-            addr = normalize_node_addr(addr_key)
-            if not addr:
-                continue
-            bucket = addresses.setdefault(addr, [])
-            if isinstance(payload, Mapping):
-                bucket.append(payload)
-            elif isinstance(payload, str):
-                bucket.append({"name": payload})
-
-    return addresses
-
-
-def _extract_snapshot_name(payloads: Iterable[Mapping[str, Any]]) -> str:
-    """Return best candidate name extracted from ``payloads``."""
-
-    queue = [payload for payload in payloads if isinstance(payload, Mapping)]
-    seen: set[int] = set()
-
-    while queue:
-        payload = queue.pop(0)
-        payload_id = id(payload)
-        if payload_id in seen:
-            continue
-        seen.add(payload_id)
-
-        for key in _SNAPSHOT_NAME_CANDIDATE_KEYS:
-            value = payload.get(key)
-            if isinstance(value, str):
-                candidate = value.strip()
-                if candidate:
-                    return candidate
-
-        queue.extend(
-            nested for nested in payload.values() if isinstance(nested, Mapping)
-        )
-
-    return ""
-
-
-
-
-def _iter_node_payload(raw_nodes: Any) -> Iterable[dict[str, Any]]:
-    """Yield node dictionaries from a payload returned by the API."""
-
-    if isinstance(raw_nodes, dict):
-        node_list = raw_nodes.get("nodes")
-        if isinstance(node_list, list):
-            for entry in node_list:
-                if isinstance(entry, dict):
-                    yield entry
-            return
-
-        seen: set[tuple[str, str]] = set()
-
-        if isinstance(node_list, Mapping):
-            yield from _iter_snapshot_sections(node_list, seen)
-
-        sections = _existing_nodes_map(raw_nodes)
-        if sections:
-            yield from _iter_snapshot_sections(sections, seen)
-            if seen:
-                return
-
-    if isinstance(raw_nodes, list):
-        for entry in raw_nodes:
-            if isinstance(entry, dict):
-                yield entry
-
-
-def _resolve_node_class(node_type: str) -> type[Node]:
-    """Return the most appropriate node class for ``node_type``."""
-
-    return NODE_CLASS_BY_TYPE.get(node_type, Node)
-
-
-def _normalise_with_fallback(
-    normalizer: Callable[..., str],
-    *candidates: Any,
-) -> str:
-    """Return the first non-empty normalised value from ``candidates``."""
-
-    for candidate in candidates:
-        normalized = normalizer(candidate, use_default_when_falsey=True)
-        if normalized:
-            return normalized
-    return ""
-
-
-def build_node_inventory(raw_nodes: Any) -> list[Node]:
-    """Return a list of :class:`Node` instances for the provided payload.
-
-    ``raw_nodes`` is typically the JSON response from ``/mgr/nodes``.  Each
-    entry is validated and normalised – unknown node types are logged at DEBUG
-    level yet still represented using the base :class:`Node` class so that
-    callers can account for the presence of the device.
-    """
-
-    inventory: list[Node] = []
-    for index, payload in enumerate(_iter_node_payload(raw_nodes)):
-        node_type = _normalise_with_fallback(
-            normalize_node_type,
-            payload.get("type"),
-            payload.get("node_type"),
-        )
-        if not node_type:
-            _LOGGER.debug("Skipping node with missing type at index %s: %s", index, payload)
-            continue
-
-        name = payload.get("name") or payload.get("title") or payload.get("label")
-        addr = _normalise_with_fallback(
-            normalize_node_addr,
-            payload.get("addr"),
-            payload.get("address"),
-        )
-
-        node_cls = _resolve_node_class(node_type)
-        if node_cls is Node:
-            _LOGGER.debug("Unsupported node type '%s' encountered", node_type)
-
-        try:
-            if node_cls is Node:
-                node = node_cls(name=name, addr=addr, node_type=node_type)
-            else:
-                node = node_cls(name=name, addr=addr)
-        except (TypeError, ValueError) as err:  # pragma: no cover - defensive
-            _LOGGER.debug("Failed to initialise node %s at index %s: %s", payload, index, err)
-            continue
-
-        inventory.append(node)
-
-    return inventory
 
 
 def ensure_node_inventory(
@@ -307,6 +76,8 @@ def ensure_node_inventory(
         mutable_record["node_inventory"] = []
 
     return []
+
+
 def parse_heater_energy_unique_id(unique_id: str) -> tuple[str, str, str] | None:
     """Parse a heater energy sensor unique ID into its components."""
 
@@ -495,8 +266,8 @@ def collect_heater_sample_addresses(
         fallback: Iterable[Any] | None = None
         if hasattr(coordinator, "_addrs"):
             try:
-                fallback = coordinator._addrs()  # type: ignore[attr-defined]
-            except Exception:  # pragma: no cover - defensive
+                fallback = coordinator._addrs()  # type: ignore[attr-defined]  # noqa: SLF001
+            except Exception:  # pragma: no cover - defensive  # noqa: BLE001
                 fallback = None
         if fallback:
             normalised: list[str] = list(normalized_map.get("htr", []))
