+++ tests/test_climate.py
@@ -25,7 +25,7 @@ from custom_components.termoweb.const import (
     signal_ws_data,
 )
 from custom_components.termoweb.inventory import HeaterNode
-from custom_components.termoweb.nodes import build_node_inventory
+from custom_components.termoweb.inventory import build_node_inventory
 from homeassistant.components.climate import HVACAction, HVACMode
 from homeassistant.const import ATTR_TEMPERATURE
 from homeassistant.core import HomeAssistant, ServiceCall
